requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.5.2
fire==0.7.0
frozenlist==1.5.0
idna==3.10
multidict==6.4.2
propcache==0.3.1
termcolor==3.0.1
typing_extensions==4.13.2
yarl==1.19.0
